+++ src/main/java/org/swyp/dessertbee/store/coupon/controller/CouponController.java
@@ -2,34 +2,24 @@ package org.swyp.dessertbee.store.coupon.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import org.swyp.dessertbee.auth.security.CustomUserDetails;
 import org.swyp.dessertbee.store.coupon.dto.request.CouponRequest;
-import org.swyp.dessertbee.store.coupon.dto.request.UseCouponRequest;
+import org.swyp.dessertbee.user.coupon.dto.request.UseCouponRequest;
 import org.swyp.dessertbee.store.coupon.dto.response.CouponResponse;
-import org.swyp.dessertbee.store.coupon.dto.response.UsedCouponResponse;
-import org.swyp.dessertbee.store.coupon.repository.CouponRepository;
-import org.swyp.dessertbee.store.coupon.scheduler.CouponScheduler;
+import org.swyp.dessertbee.user.coupon.dto.response.UsedCouponResponse;
 import org.swyp.dessertbee.store.coupon.service.CouponService;
-import org.swyp.dessertbee.store.coupon.service.UserCouponService;
+import org.swyp.dessertbee.user.coupon.service.UserCouponService;
 import org.swyp.dessertbee.store.store.entity.Store;
 import org.swyp.dessertbee.store.store.repository.StoreRepository;
-import org.swyp.dessertbee.user.entity.UserEntity;
-import org.swyp.dessertbee.user.service.UserService;
 
 import java.util.List;
-import java.util.UUID;
 
 @RestController
 @RequestMapping("/api/coupons")
 @RequiredArgsConstructor
 public class CouponController {
-    //todo: Role 부여 - 사장님만 가능하게
 
     private final CouponService couponService;
     private final UserCouponService userCouponService;
@@ -38,6 +28,7 @@ public class CouponController {
     /**
      * 쿠폰 생성 (1인 1쿠폰)
      */
+    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_OWNER', 'ROLE_ADMIN')")
     @PostMapping("/create")
     public ResponseEntity<CouponResponse> createCoupon(@RequestBody CouponRequest request) {
         Store store = storeRepository.findByStoreUuid(request.getStoreUuid())
@@ -47,6 +38,7 @@ public class CouponController {
     /**
      * 쿠폰 수정
      */
+    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_OWNER', 'ROLE_ADMIN')")
     @PutMapping("/{couponId}")
     public ResponseEntity<CouponResponse> updateCoupon(
             @PathVariable Long couponId,
@@ -61,6 +53,7 @@ public class CouponController {
     /**
      * 쿠폰 삭제
      */
+    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_OWNER', 'ROLE_ADMIN')")
     @DeleteMapping("/{couponId}")
     public ResponseEntity<Void> deleteCoupon(
             @PathVariable Long couponId
@@ -72,6 +65,7 @@ public class CouponController {
     /**
      * 생성한 쿠폰 조회
      */
+    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_OWNER', 'ROLE_ADMIN')")
     @GetMapping("/all")
     public ResponseEntity<List<CouponResponse>> getAllCoupons() {
         return ResponseEntity.ok(couponService.getAllCoupons());
@@ -80,6 +74,7 @@ public class CouponController {
     /**
      * 쿠폰 사용처리
      */
+    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_OWNER', 'ROLE_ADMIN')")
     @PostMapping("/use")
     public ResponseEntity<UsedCouponResponse> useCoupon(@RequestBody UseCouponRequest request) {
         UsedCouponResponse response = userCouponService.useCouponByCode(request);
