+++ src/main/java/org/swyp/dessertbee/user/coupon/dto/response/UserCouponDetailResponse.java
@@ -1,7 +1,6 @@
-package org.swyp.dessertbee.store.coupon.dto.response;
+package org.swyp.dessertbee.user.coupon.dto.response;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 import org.swyp.dessertbee.store.coupon.entity.enums.CouponConditionType;
